Added key bindings. Fixed some memory leaks.

+++ src/GLee/GLee.c
@@ -15913,7 +15913,7 @@ void __GLeeExtList_add(ExtensionList *extList, const char * extName)
 	int n=i+1;
 	if (i==0)
 	{
-		extList->lengths=(size_t *)malloc(sizeof(int));
+		extList->lengths=(size_t *)malloc(sizeof(size_t));
 		extList->names=(char **)malloc(sizeof(char *));
 	}else
 	{
